# mw-general-bff/apiAutogenerated/mw-general-bff/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  title: Masters way general-bff API
  version: "1.0"
servers:
- url: /general
paths:
...
  /auth/login/local/{userEmail}:
    get:
      description: Login locally by providing an email address.
      operationId: get token locally
      parameters:
      - description: email
        in: path
        name: userEmail
        required: true
        schema:
          type: string
      responses:
        "304":
          content: {}
          description: redirect
      summary: login locally by email (with no oauth)
      tags:
      - auth
  /auth/logout/{provider}:
    get:
      operationId: logout-current-authorized-user
      parameters:
      - description: google
        in: path
        name: provider
        required: true
        schema:
          type: string
      responses:
        "204":
          content: {}
          description: No Content
      summary: Logout current authorized user
      tags:
      - auth
  /auth/refreshToken:
    post:
      operationId: refresh-access-token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mw-general-bff_internal_schemas.RefreshAccessTokenPayload'
        description: query params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.RefreshAccessTokenResponse'
          description: OK
      summary: Retrieve Access Token
      tags:
      - auth
      x-codegen-request-body-name: request
  /auth/{provider}:
    get:
      operationId: begin-auth
      parameters:
      - description: google
        in: path
        name: provider
        required: true
        schema:
          type: string
      responses:
        "307":
          content: {}
          description: Temporary Redirect
      summary: Begin oauth
      tags:
      - auth
  /auth/{provider}/callback:
    get:
      operationId: google auth log in callback function
      parameters:
      - description: state parameter
        in: query
        name: state
        required: true
        schema:
          type: string
      - description: google
        in: path
        name: provider
        required: true
        schema:
          type: string
      responses:
        "302":
          content:
            application/json:
              schema:
                type: string
          description: Redirect to frontend with JWT token
      summary: Log in with google oAuth
      tags:
      - auth
...
  /jobDones:
    post:
      operationId: create-jobDone
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mw-general-bff_internal_schemas.CreateJobDonePayload'
        description: query params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.JobDonePopulatedResponse'
          description: OK
      summary: Create a new jobDone
      tags:
      - jobDone
      x-codegen-request-body-name: request
  /jobDones/{jobDoneId}:
    delete:
      operationId: delete-jobDone
      parameters:
      - description: jobDone ID
        in: path
        name: jobDoneId
        required: true
        schema:
          type: string
      responses:
        "204":
          content: {}
          description: No Content
      summary: Delete jobDone by UUID
      tags:
      - jobDone
    patch:
      operationId: update-jobDone
      parameters:
      - description: jobDone UUID
        in: path
        name: jobDoneId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mw-general-bff_internal_schemas.UpdateJobDone'
        description: query params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.JobDonePopulatedResponse'
          description: OK
      summary: Update jobDone by UUID
      tags:
      - jobDone
      x-codegen-request-body-name: request
...
components:
  schemas:
...
    mw-general-bff_internal_schemas.CreateJobDonePayload:
      properties:
        dayReportUuid:
          type: string
        description:
          type: string
        jobTagUuids:
          items:
            type: string
          type: array
        ownerUuid:
          type: string
        time:
          type: integer
      required:
      - dayReportUuid
      - description
      - jobTagUuids
      - ownerUuid
      - time
      type: object
...
    mw-general-bff_internal_schemas.JobDonePopulatedResponse:
      example:
        createdAt: createdAt
        ownerUuid: ownerUuid
        wayName: wayName
        ownerName: ownerName
        wayUuid: wayUuid
        description: description
        time: 0
        dayReportUuid: dayReportUuid
        uuid: uuid
        tags:
        - color: color
          name: name
          description: description
          uuid: uuid
        - color: color
          name: name
          description: description
          uuid: uuid
        updatedAt: updatedAt
      properties:
        createdAt:
          type: string
        dayReportUuid:
          type: string
        description:
          type: string
        ownerName:
          type: string
        ownerUuid:
          type: string
        tags:
          items:
            $ref: '#/components/schemas/mw-general-bff_internal_schemas.JobTagResponse'
          type: array
        time:
          type: integer
        updatedAt:
          type: string
        uuid:
          type: string
        wayName:
          type: string
        wayUuid:
          type: string
      required:
      - createdAt
      - dayReportUuid
      - description
      - ownerName
      - ownerUuid
      - tags
      - time
      - updatedAt
      - uuid
      - wayName
      - wayUuid
      type: object
    mw-general-bff_internal_schemas.JobTagResponse:
      example:
        color: color
        name: name
        description: description
        uuid: uuid
      properties:
        color:
          type: string
        description:
          type: string
        name:
          type: string
        uuid:
          type: string
      required:
      - color
      - description
      - name
      - uuid
      type: object
...
    mw-general-bff_internal_schemas.RefreshAccessTokenPayload:
      properties:
        refreshToken:
          type: string
      required:
      - refreshToken
      type: object
    mw-general-bff_internal_schemas.RefreshAccessTokenResponse:
      example:
        accessToken: accessToken
      properties:
        accessToken:
          type: string
      required:
      - accessToken
      type: object
...
    mw-general-bff_internal_schemas.UpdateJobDone:
      properties:
        description:
          type: string
        time:
          type: integer
      type: object
...