# mw-general-bff/internal/controllers/auth.controller.go
package controllers

import (
	openapiGeneral "mw-general-bff/apiAutogenerated/general"
	"mw-general-bff/internal/auth"
	"mw-general-bff/internal/facades"
	"mw-general-bff/internal/schemas"
	"mw-general-bff/pkg/utils"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/samber/lo"
)

// Without next lines swagger does not see openapi models
var _ = &openapiGeneral.MwServerInternalSchemasUserPopulatedResponse{}

type AuthController struct {
	authFacade *facades.AuthFacade
}

func NewAuthController(authFacade *facades.AuthFacade) *AuthController {
	return &AuthController{authFacade}
}

// Log in with google oAuth
// @Summary Log in with google oAuth
// @Description
// @Tags auth
// @ID google auth log in callback function
// @Accept  json
// @Produce  json
// @Param state query string true "state parameter"
// @Param provider path string true "google"
// @Success 302 {string} string "Redirect to frontend with JWT token"
// @Router /auth/{provider}/callback [get]
func (ac *AuthController) GetAuthCallbackFunction(ctx *gin.Context) {
	provider := ctx.Param("provider")

	code := ctx.Query("code")
	state := ctx.Query("state")
	if state != auth.OauthStateString {
		ctx.JSON(http.StatusUnauthorized, gin.H{"error": "invalid oauth state"})
		return
	}

	authCallback, err := ac.authFacade.GetAuthCallbackFunction(ctx, provider, code, state)
	utils.HandleErrorGin(ctx, err)

	ctx.Redirect(http.StatusFound, authCallback.Url)
}

// Begin auth handler
// @Summary Begin oauth
// @Description
// @Tags auth
// @ID begin-auth
// @Accept  json
// @Produce  json
// @Param provider path string true "google"
// @Success 307
// @Router /auth/{provider} [get]
func (ac *AuthController) BeginAuth(ctx *gin.Context) {
	provider := ctx.Param("provider")

	response, err := ac.authFacade.GetGoogleAuthURL(ctx, provider)
	utils.HandleErrorGin(ctx, err)

	ctx.Redirect(http.StatusTemporaryRedirect, response.Url)
}

// @Summary Get current authorized user
// @Description
// @Tags auth
// @ID get-current-authorized-user
// @Accept  json
// @Produce  json
// @Success 200 {object} schemas.UserPopulatedResponse
// @Router /auth/current [get]
func (ac *AuthController) GetCurrentAuthorizedUserByToken(ctx *gin.Context) {
	populatedUser, err := ac.authFacade.GetCurrentAuthorizedUserByToken(ctx)
	utils.HandleErrorGin(ctx, err)

	customWayCollections := lo.Map(populatedUser.CustomWayCollections, func(collection openapiGeneral.MwServerInternalSchemasWayCollectionPopulatedResponse, _ int) schemas.WayCollectionPopulatedResponse {
		return schemas.WayCollectionPopulatedResponse{
			Uuid: collection.Uuid,
			Name: collection.Name,
			Ways: lo.Map(collection.Ways, func(way openapiGeneral.MwServerInternalSchemasWayPlainResponse, _ int) schemas.WayPlainResponse {
				return schemas.WayPlainResponse{
					Uuid:            way.Uuid,
					Name:            way.Name,
					GoalDescription: way.GoalDescription,
					UpdatedAt:       way.UpdatedAt,
					CreatedAt:       way.CreatedAt,
					EstimationTime:  way.EstimationTime,
					IsCompleted:     way.IsCompleted,
					Owner: schemas.UserPlainResponse{
						Uuid:        way.Owner.Uuid,
						Name:        way.Owner.Name,
						Email:       way.Owner.Email,
						Description: way.Owner.Description,
						CreatedAt:   way.Owner.CreatedAt,
						ImageUrl:    way.Owner.ImageUrl,
						IsMentor:    way.Owner.IsMentor,
					},
					CopiedFromWayUuid: way.CopiedFromWayUuid.Get(),
					ProjectUuid:       way.ProjectUuid.Get(),
					IsPrivate:         way.IsPrivate,
					FavoriteForUsers:  way.FavoriteForUsers,
					DayReportsAmount:  way.DayReportsAmount,
					Mentors: lo.Map(way.Mentors, func(mentor openapiGeneral.MwServerInternalSchemasUserPlainResponse, _ int) schemas.UserPlainResponse {
						return schemas.UserPlainResponse{
							Uuid:        mentor.Uuid,
							Name:        mentor.Name,
							Email:       mentor.Email,
							Description: mentor.Description,
							CreatedAt:   mentor.CreatedAt,
							ImageUrl:    mentor.ImageUrl,
							IsMentor:    mentor.IsMentor,
						}
					}),
					MetricsDone:   way.MetricsDone,
					MetricsTotal:  way.MetricsTotal,
					ChildrenUuids: way.ChildrenUuids,
					WayTags: lo.Map(way.WayTags, func(tag openapiGeneral.MwServerInternalSchemasWayTagResponse, _ int) schemas.WayTagResponse {
						return schemas.WayTagResponse{
							Uuid: tag.Uuid,
							Name: tag.Name,
						}
					}),
				}
			}),
			CreatedAt: collection.CreatedAt,
			UpdatedAt: collection.UpdatedAt,
			OwnerUuid: collection.OwnerUuid,
			Type:      collection.Type,
		}
	})

	defaultWayCollections := schemas.DefaultWayCollections{
		Own: schemas.WayCollectionPopulatedResponse{
			Uuid:      populatedUser.DefaultWayCollections.Own.Uuid,
			Name:      populatedUser.DefaultWayCollections.Own.Name,
			CreatedAt: populatedUser.DefaultWayCollections.Own.CreatedAt,
			UpdatedAt: populatedUser.DefaultWayCollections.Own.UpdatedAt,
			OwnerUuid: populatedUser.DefaultWayCollections.Own.OwnerUuid,
			Type:      populatedUser.DefaultWayCollections.Own.Type,
			Ways:      lo.Map(populatedUser.DefaultWayCollections.Own.Ways, ConvertWay),
		},
		Favorite: schemas.WayCollectionPopulatedResponse{
			Uuid:      populatedUser.DefaultWayCollections.Own.Uuid,
			Name:      populatedUser.DefaultWayCollections.Own.Name,
			CreatedAt: populatedUser.DefaultWayCollections.Own.CreatedAt,
			UpdatedAt: populatedUser.DefaultWayCollections.Own.UpdatedAt,
			OwnerUuid: populatedUser.DefaultWayCollections.Own.OwnerUuid,
			Type:      populatedUser.DefaultWayCollections.Own.Type,
			Ways:      lo.Map(populatedUser.DefaultWayCollections.Own.Ways, ConvertWay),
		},
		Mentoring: schemas.WayCollectionPopulatedResponse{
			Uuid:      populatedUser.DefaultWayCollections.Own.Uuid,
			Name:      populatedUser.DefaultWayCollections.Own.Name,
			CreatedAt: populatedUser.DefaultWayCollections.Own.CreatedAt,
			UpdatedAt: populatedUser.DefaultWayCollections.Own.UpdatedAt,
			OwnerUuid: populatedUser.DefaultWayCollections.Own.OwnerUuid,
			Type:      populatedUser.DefaultWayCollections.Own.Type,
			Ways:      lo.Map(populatedUser.DefaultWayCollections.Own.Ways, ConvertWay),
		},
	}

	response := schemas.UserPopulatedResponse{
		Uuid:               populatedUser.Uuid,
		Name:               populatedUser.Name,
		Email:              populatedUser.Email,
		Description:        populatedUser.Description,
		CreatedAt:          populatedUser.CreatedAt,
		ImageUrl:           populatedUser.ImageUrl,
		IsMentor:           populatedUser.IsMentor,
		WayCollections:     customWayCollections,
		DefaultCollections: defaultWayCollections,
		FavoriteForUsers:   populatedUser.FavoriteForUsers,
		FavoriteUsers: lo.Map(populatedUser.FavoriteUsers, func(user openapiGeneral.MwServerInternalSchemasUserPlainResponse, _ int) schemas.UserPlainResponse {
			return schemas.UserPlainResponse{
				Uuid:        user.Uuid,
				Name:        user.Name,
				Email:       user.Email,
				Description: user.Description,
				CreatedAt:   user.CreatedAt,
				ImageUrl:    user.ImageUrl,
				IsMentor:    user.IsMentor,
			}
		}),
		Tags: lo.Map(populatedUser.Tags, func(tag openapiGeneral.MwServerInternalSchemasUserTagResponse, _ int) schemas.UserTagResponse {
			return schemas.UserTagResponse{
				Uuid: tag.Uuid,
				Name: tag.Name,
			}
		}),
		WayRequests: lo.Map(populatedUser.WayRequests, ConvertWay),
		Projects: lo.Map(populatedUser.Projects, func(project openapiGeneral.MwServerInternalSchemasProjectPlainResponse, _ int) schemas.ProjectPlainResponse {
			return schemas.ProjectPlainResponse{
				ID:        project.Id,
				Name:      project.Name,
				IsPrivate: project.IsPrivate,
				UserIDs:   project.UserIds,
			}
		}),
	}

	ctx.JSON(http.StatusOK, response)
}

// GetUserTokenByEmail handler
// @Summary login locally by email (with no oauth)
// @Description Login locally by providing an email address.
// @Tags auth
// @ID get token locally
// @Accept  json
// @Produce  json
// @Param userEmail path string true "email"
// @Success 304 "redirect"
// @Router /auth/login/local/{userEmail} [get]
func (ac *AuthController) GetUserTokenByEmail(ctx *gin.Context) {
	userEmail := ctx.Param("userEmail")

	authCallback, err := ac.authFacade.GetUserTokenByEmail(ctx, userEmail)
	utils.HandleErrorGin(ctx, err)

	ctx.Redirect(http.StatusFound, authCallback.Url)
}

// @Summary Logout current authorized user
// @Description
// @Tags auth
// @ID logout-current-authorized-user
// @Accept  json
// @Produce  json
// @Param provider path string true "google"
// @Success 204
// @Router /auth/logout/{provider} [get]
func (ac *AuthController) Logout(ctx *gin.Context) {
	provider := ctx.Param("provider")

	err := ac.authFacade.Logout(ctx, provider)
	utils.HandleErrorGin(ctx, err)

	ctx.Status(http.StatusNoContent)
}

// @Summary Retrieve Google Access Token
// @Description This endpoint retrieves the Google access token for an authenticated user.
// @Tags auth
// @ID get-google-token
// @Accept json
// @Produce json
// @Success 200 {object} openapiGeneral.MwServerInternalSchemasGoogleToken
// @Router /auth/google-token [get]
func (ac *AuthController) GetGoogleAccessToken(ctx *gin.Context) {
	token, err := ac.authFacade.GetGoogleAccessTokenByUserID(ctx)
	utils.HandleErrorGin(ctx, err)

	ctx.JSON(http.StatusOK, token)
}

// @Summary Retrieve Access Token
// @Description
// @Tags auth
// @ID refresh-access-token
// @Accept json
// @Produce json
// @Param request body schemas.RefreshAccessTokenPayload true "query params"
// @Success 200 {object} schemas.RefreshAccessTokenResponse
// @Router /auth/refreshToken [post]
func (ac *AuthController) RefreshAccessToken(ctx *gin.Context) {
	var payload *schemas.RefreshAccessTokenPayload

	if err := ctx.ShouldBindJSON(&payload); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"status": "Failed payload", "error": err.Error()})
		return
	}

	responseRaw, err := ac.authFacade.RefreshAccessToken(ctx, payload.RefreshToken)
	utils.HandleErrorGin(ctx, err)

	response := schemas.RefreshAccessTokenResponse{
		AccessToken: responseRaw.AccessToken,
	}

	ctx.JSON(http.StatusOK, response)
}

func ConvertWay(way openapiGeneral.MwServerInternalSchemasWayPlainResponse, _ int) schemas.WayPlainResponse {
	return schemas.WayPlainResponse{
		Uuid:            way.Uuid,
		Name:            way.Name,
		GoalDescription: way.GoalDescription,
		UpdatedAt:       way.UpdatedAt,
		CreatedAt:       way.CreatedAt,
		EstimationTime:  way.EstimationTime,
		IsCompleted:     way.IsCompleted,
		Owner: schemas.UserPlainResponse{
			Uuid:        way.Owner.Uuid,
			Name:        way.Owner.Name,
			Email:       way.Owner.Email,
			Description: way.Owner.Description,
			CreatedAt:   way.Owner.CreatedAt,
			ImageUrl:    way.Owner.ImageUrl,
			IsMentor:    way.Owner.IsMentor,
		},
		CopiedFromWayUuid: way.CopiedFromWayUuid.Get(),
		ProjectUuid:       way.ProjectUuid.Get(),
		IsPrivate:         way.IsPrivate,
		FavoriteForUsers:  way.FavoriteForUsers,
		DayReportsAmount:  way.DayReportsAmount,
		Mentors: lo.Map(way.Mentors, func(mentor openapiGeneral.MwServerInternalSchemasUserPlainResponse, _ int) schemas.UserPlainResponse {
			return schemas.UserPlainResponse{
				Uuid:        mentor.Uuid,
				Name:        mentor.Name,
				Email:       mentor.Email,
				Description: mentor.Description,
				CreatedAt:   mentor.CreatedAt,
				ImageUrl:    mentor.ImageUrl,
				IsMentor:    mentor.IsMentor,
			}
		}),
		MetricsDone:   way.MetricsDone,
		MetricsTotal:  way.MetricsTotal,
		ChildrenUuids: way.ChildrenUuids,
		WayTags: lo.Map(way.WayTags, func(tag openapiGeneral.MwServerInternalSchemasWayTagResponse, _ int) schemas.WayTagResponse {
			return schemas.WayTagResponse{
				Uuid: tag.Uuid,
				Name: tag.Name,
			}
		}),
	}
}