# mw-general-bff/internal/controllers/jobDone.controller.go
package controllers

import (
	openapiGeneral "mw-general-bff/apiAutogenerated/general"
	"mw-general-bff/internal/auth"
	"mw-general-bff/internal/facades"
	"mw-general-bff/internal/schemas"
	"mw-general-bff/internal/services"
	"mw-general-bff/pkg/utils"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/samber/lo"
)

// Without next lines swagger does not see openapi models
var _ = &openapiGeneral.MwServerInternalSchemasJobDonePopulatedResponse{}

type JobDoneController struct {
	jobDoneFacade *facades.JobDoneFacade
}

func NewJobDoneController(jobDoneFacade *facades.JobDoneFacade) *JobDoneController {
	return &JobDoneController{jobDoneFacade}
}

// Create JobDone handler
// @Summary Create a new jobDone
// @Description
// @Tags jobDone
// @ID create-jobDone
// @Accept  json
// @Produce  json
// @Param request body schemas.CreateJobDonePayload true "query params"
// @Success 200 {object} schemas.JobDonePopulatedResponse
// @Router /jobDones [post]
func (jc *JobDoneController) CreateJobDone(ctx *gin.Context) {
	var payload *schemas.CreateJobDonePayload

	if err := ctx.ShouldBindJSON(&payload); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"status": "Failed payload", "error": err.Error()})
		return
	}

	responseRaw, err := jc.jobDoneFacade.CreateJobDone(ctx, payload)
	utils.HandleErrorGin(ctx, err)

	response := schemas.JobDonePopulatedResponse{
		Uuid:          responseRaw.Uuid,
		CreatedAt:     responseRaw.CreatedAt,
		UpdatedAt:     responseRaw.UpdatedAt,
		Description:   responseRaw.Description,
		Time:          responseRaw.Time,
		OwnerUuid:     responseRaw.OwnerUuid,
		OwnerName:     responseRaw.OwnerName,
		DayReportUuid: responseRaw.DayReportUuid,
		WayUUID:       responseRaw.WayUuid,
		WayName:       responseRaw.WayName,
		Tags: lo.Map(responseRaw.Tags, func(tag openapiGeneral.MwServerInternalSchemasJobTagResponse, _ int) schemas.JobTagResponse {
			return schemas.JobTagResponse{
				Uuid:        tag.Uuid,
				Name:        tag.Name,
				Description: tag.Description,
				Color:       tag.Color,
			}
		}),
	}

	ctx.JSON(http.StatusOK, response)
}

// Update JobDone handler
// @Summary Update jobDone by UUID
// @Description
// @Tags jobDone
// @ID update-jobDone
// @Accept  json
// @Produce  json
// @Param request body schemas.UpdateJobDone true "query params"
// @Param jobDoneId path string true "jobDone UUID"
// @Success 200 {object} schemas.JobDonePopulatedResponse
// @Router /jobDones/{jobDoneId} [patch]
func (jc *JobDoneController) UpdateJobDone(ctx *gin.Context) {
	var payload *schemas.UpdateJobDone
	jobDoneID := ctx.Param("jobDoneId")

	modifierUserIdRaw, _ := ctx.Get(auth.ContextKeyUserID)
	modifierUserId := modifierUserIdRaw.(string)

	if err := ctx.ShouldBindJSON(&payload); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"status": "Failed payload", "error": err.Error()})
		return
	}

	params := &services.UpdateJobDoneParams{
		JobDoneID:        jobDoneID,
		Description:      payload.Description,
		Time:             payload.Time,
		ModifierUserUuid: modifierUserId,
	}

	responseRaw, err := jc.jobDoneFacade.UpdateJobDone(ctx, params)
	utils.HandleErrorGin(ctx, err)

	response := schemas.JobDonePopulatedResponse{
		Uuid:          responseRaw.Uuid,
		CreatedAt:     responseRaw.CreatedAt,
		UpdatedAt:     responseRaw.UpdatedAt,
		Description:   responseRaw.Description,
		Time:          responseRaw.Time,
		OwnerUuid:     responseRaw.OwnerUuid,
		OwnerName:     responseRaw.OwnerName,
		DayReportUuid: responseRaw.DayReportUuid,
		WayUUID:       responseRaw.WayUuid,
		WayName:       responseRaw.WayName,
		Tags: lo.Map(responseRaw.Tags, func(tag openapiGeneral.MwServerInternalSchemasJobTagResponse, _ int) schemas.JobTagResponse {
			return schemas.JobTagResponse{
				Uuid:        tag.Uuid,
				Name:        tag.Name,
				Description: tag.Description,
				Color:       tag.Color,
			}
		}),
	}

	ctx.JSON(http.StatusOK, response)
}

// Deleting jobDone handler
// @Summary Delete jobDone by UUID
// @Description
// @Tags jobDone
// @ID delete-jobDone
// @Accept  json
// @Produce  json
// @Param jobDoneId path string true "jobDone ID"
// @Success 204
// @Router /jobDones/{jobDoneId} [delete]
func (jc *JobDoneController) DeleteJobDoneById(ctx *gin.Context) {
	jobDoneID := ctx.Param("jobDoneId")

	err := jc.jobDoneFacade.DeleteJobDoneByID(ctx, jobDoneID)
	utils.HandleErrorGin(ctx, err)

	ctx.Status(http.StatusNoContent)
}